ref: improve the package structure
