+++ airbyte-ci/connectors/base_images/CHANGELOG_PYTHON_CONNECTOR_BASE_IMAGE.md
@@ -1,7 +1,7 @@
 # Changelog for airbyte-python-connector-base
 
-|                                                        Version                                                        |                                                           Changelog                                                          |
-|-----------------------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------------------------------------------------------------|
-|[1.1.1](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/base_images/base_images/python_bases.py)|                                       Fix: we should use Poetry 1.6.0 instead of 1.6.1                                       |
-|[1.1.0](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/base_images/base_images/python_bases.py)|                                                     Install poetry 1.6.1                                                     |
-|[1.0.0](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/base_images/base_images/python_bases.py)|Declare our first base image version. It uses Python 3.9.18 on a Debian 11 (Bookworm) system with Pip 23.2.1 and UTC timezone.|
+|                                                       Version                                                      |                                                           Changelog                                                          |
+|--------------------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------------------------------------------------------------|
+|[1.1.1](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/base_images/base_images/python/v1.py)|                                       Fix: we should use Poetry 1.6.0 instead of 1.6.1                                       |
+|[1.1.0](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/base_images/base_images/python/v1.py)|                                                     Install poetry 1.6.1                                                     |
+|[1.0.0](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/base_images/base_images/python/v1.py)|Declare our first base image version. It uses Python 3.9.18 on a Debian 11 (Bookworm) system with Pip 23.2.1 and UTC timezone.|
